Trait attestation_agent::AttestationAPIs

pub trait AttestationAPIs {
    // Required methods
    fn decrypt_image_layer_annotation<'life0, 'life1, 'life2, 'life3, 'async_trait>(
        &'life0 mut self,
        kbc_name: &'life1 str,
        kbs_uri: &'life2 str,
        annotation: &'life3 str
    ) -> Pin<Box<dyn Future<Output = Result<Vec<u8>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             'life3: 'async_trait;
    fn download_confidential_resource<'life0, 'life1, 'life2, 'life3, 'async_trait>(
        &'life0 mut self,
        kbc_name: &'life1 str,
        resource_path: &'life2 str,
        kbs_uri: &'life3 str
    ) -> Pin<Box<dyn Future<Output = Result<Vec<u8>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             'life3: 'async_trait;
    fn get_token<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        token_type: &'life1 str
    ) -> Pin<Box<dyn Future<Output = Result<Vec<u8>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn get_evidence<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        runtime_data: &'life1 [u8]
    ) -> Pin<Box<dyn Future<Output = Result<Vec<u8>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
}

Attestation Agent (AA for short) is a rust library crate for attestation procedure in confidential containers. It provides kinds of service APIs that need to make requests to the Relying Party (Key Broker Service) in Confidential Containers, and establishes an attestation and connection between the corresponding Key Broker Client (KBC) and KBS, so as to obtain the trusted services or resources of KBS.

Example

use attestation_agent::AttestationAgent;
use attestation_agent::AttestationAPIs;

let mut aa = AttestationAgent::new();

let key_result = aa.decrypt_image_layer_annotation(
    "sample_kbc",
    "https://xxxxx",
    "example_annotation"
);

AttestationAPIs defines the service APIs of attestation agent that need to make requests to the Relying Party (Key Broker Service) in Confidential Containers.

For every service API, the kbc_name and kbs_uri is necessary, kbc_name tells attestation agent which KBC module it should use and kbs_uri specifies the KBS address.

Required Methods

fn decrypt_image_layer_annotation<'life0, 'life1, 'life2, 'life3, 'async_trait>( &'life0 mut self, kbc_name: &'life1 str, kbs_uri: &'life2 str, annotation: &'life3 str ) -> Pin<Box<dyn Future<Output = Result<Vec<u8>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, 'life3: 'async_trait,

Decrypt the encrypted information in annotation.

The specific format of annotation is defined by different KBC and corresponding KBS. The decryption method may be to obtain the key from KBS for decryption, or directly send the annotation to KBS for decryption, which depends on the specific implementation of each KBC module.

TODO: move this API to Confidential Data Hub

fn download_confidential_resource<'life0, 'life1, 'life2, 'life3, 'async_trait>( &'life0 mut self, kbc_name: &'life1 str, resource_path: &'life2 str, kbs_uri: &'life3 str ) -> Pin<Box<dyn Future<Output = Result<Vec<u8>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, 'life3: 'async_trait,

Request KBS to obtain confidential resources, including confidential data or files.

resource_uri is a KBS Resource URI pointing to a specific resource.

TODO: remove this API

fn get_token<'life0, 'life1, 'async_trait>( &'life0 mut self, token_type: &'life1 str ) -> Pin<Box<dyn Future<Output = Result<Vec<u8>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Get attestation Token

fn get_evidence<'life0, 'life1, 'async_trait>( &'life0 mut self, runtime_data: &'life1 [u8] ) -> Pin<Box<dyn Future<Output = Result<Vec<u8>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Get TEE hardware signed evidence that includes the runtime data.